mello-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
cryptography==41.0.7
firebase-admin==6.2.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
